✨ feat: add data tools, binary packing, and nnue embed
Add command-line tools for dataset processing (converter, epdproc,
shuffler, sprt) and Makefile targets to build and embed networks.
Introduce a compact 32-byte PackedPos binary format with serialization
and tests to enable efficient training I/O and shuffling.
Add perft helper, integrate embedded NNUE via go:embed, and expose
network info in UCI. Adjust SIMD evaluation quantization and update
docs and lichess config for dataset/training workflows.

+++ internal/nnue/simd_x64.go
@@ -17,7 +17,7 @@ func updateAccumulatorAVX2(acc *types.Accumulator, weights *int16)
 func subAccumulatorAVX2(acc *types.Accumulator, weights *int16)
 
 //go:noescape
-func evaluateAVX2(us, them *types.Accumulator, weights *int16, bias int32) int32
+func evaluateAVX2(us, them *types.Accumulator, weights *int16, bias int16) int32
 
 var (
 	hasAVX2 = false
